# api/spec/static/api/openapi/node/v1alpha1/openapi.yaml
...
openapi: 3.0.3
info:
    title: Identity Node
    description: Identity Node
    contact:
        name: Agntcy Team
        url: https://github.com/agntcy/identity
        email: identity@agntcy.org
    version: v1alpha1
servers:
    - url: http://0.0.0.0:4000
      description: Local environment
paths:
...
    /v1alpha1/vc/publish:
        post:
            tags:
                - VcService
            description: Publish an issued Verifiable Credential
            operationId: VcService_Publish
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PublishRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1alpha1/vc/search:
        post:
            tags:
                - VcService
            description: Search for Verifiable Credentials based on the specified criteria
            operationId: VcService_Search
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SearchRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1alpha1/vc/verify:
        post:
            tags:
                - VcService
            description: Verify an existing Verifiable Credential
            operationId: VcService_Verify
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VerifyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1alpha1/vc/{id}/.well-known/vcs.json:
        get:
            tags:
                - VcService
            description: Returns the well-known Verifiable Credentials for the specified Id
            operationId: VcService_GetWellKnown
            parameters:
                - name: id
                  in: path
                  description: The identifier.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetVcWellKnownResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CredentialSchema:
            type: object
            properties:
                type:
                    type: string
                    description: Type specifies the type of the file
                id:
                    type: string
                    description: The URL identifying the schema file
            description: |-
                CredentialSchema represents the credentialSchema property of a Verifiable Credential.
                 more information can be found [here]

                 [here]: https://www.w3.org/TR/vc-data-model-2.0/#data-schemas
        EnvelopedCredential:
            type: object
            properties:
                envelopeType:
                    enum:
                        - CREDENTIAL_ENVELOPE_TYPE_UNSPECIFIED
                        - CREDENTIAL_ENVELOPE_TYPE_EMBEDDED_PROOF
                        - CREDENTIAL_ENVELOPE_TYPE_JOSE
                    type: string
                    description: EnvelopeType specifies the type of the envelope used to store the credential.
                    format: enum
                value:
                    type: string
                    description: Value is the enveloped credential in the specified format.
            description: EnvelopedCredential represents a Credential enveloped in a specific format.
...
        GetVcWellKnownResponse:
            type: object
            properties:
                vcs:
                    type: array
                    items:
                        $ref: '#/components/schemas/EnvelopedCredential'
                    description: The list of VCs for the specified Id
            description: Returns the well-known VCs for an Id
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        Proof:
            type: object
            properties:
                type:
                    type: string
                    description: The type of the proof
                proofPurpose:
                    type: string
                    description: The proof purpose
                proofValue:
                    type: string
                    description: The proof value
            description: |-
                A data integrity proof provides information about the proof mechanism,
                 parameters required to verify that proof, and the proof value itself.
        PublishRequest:
            type: object
            properties:
                vc:
                    allOf:
                        - $ref: '#/components/schemas/EnvelopedCredential'
                    description: The Verifiable Credential to publish
                proof:
                    allOf:
                        - $ref: '#/components/schemas/Proof'
                    description: |-
                        Optional Proof of ownership of the Issuer's ResolverMetadata
                         This should be provided when the Issuer is provided by an external IdP
                         Example: a signed JWT
            description: Request to publish an issued Verifiable Credential
...
        SearchRequest:
            type: object
            properties:
                id:
                    type: string
                    description: ID is the identifier.
                schema:
                    allOf:
                        - $ref: '#/components/schemas/CredentialSchema'
                    description: Schema is the schema of the credential contained in the VC
                content:
                    type: string
                    description: The credential content representation in JSON-LD format
            description: Request to search for VCs based on the specified criteria
        SearchResponse:
            type: object
            properties:
                vcs:
                    type: array
                    items:
                        $ref: '#/components/schemas/EnvelopedCredential'
                    description: The list of VCs that match the search criteria
            description: Returns the VCs that match the search criteria
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        VerifyRequest:
            type: object
            properties:
                vc:
                    allOf:
                        - $ref: '#/components/schemas/EnvelopedCredential'
                    description: The Verifiable Credential to verify
            description: Request to verify an existing Verifiable Credential
...
    securitySchemes: {}
security:
    - {}
...